all changes many files to use new mobile view for charts
offers chart use when viewed in landscape view on mobile devices.note almanac data is disavled in mobile view these need reworking to be able to use clearly in mobile landscape view. this is quite a substantial change that has been implemented in weather34.com for some time. this was not planned I had seen Williams attempt and thought i do have that to hand, i hope this does not cause any inconvience or confusion. brian