Rewriting mb.php
**THIS UPDATE REQUIRES A REBOOT OF YOUR METEOBRIDGE**
This is a fairly big update as far as this file is concerned. mb.php will now only save to the database if 5 minutes have passed.
The newly updated API includes 5 min averages so that no data is lost during these times.
The goal of this is to smooth out the daily charts, while not losing any data

Also includes a fix for your advisory box in positon 4 not updating right