Add optional video stream to webcam popup...
This is a multi-part update

First, added an optional video stream to the live webcam popup to actually make it live.
Added new setting to easyweathersetup.php to input your video link alongside the existing static image link.
If your particular video link doesn't work, but you can view it in a normal browser with no extra extensions or anything, submit an issue on GitHub or email me and I'll attempt to add support for your live webcam.
If you don't want this, just leave the video link empty and the popup will use the static image link instead.

Secondly, a fairly big formatting and alignment update to the easyweathersetup.php
I tried to fix the odd alignment of the sections and break up some of the larger sections that covered multiple topics.
Fixed a bit of the formatting across all the descriptive texts on the page.