Linking to the World...
Added a bunch of optional links, several that are predefined and just need a username/id, and 3 that are completely customizable (provide the name and url).
The flyout menu for the extra links has a setting to enable/disable it, and each individual link is automatically hidden when the link's information isn't filled in.
Added svg logos for all of the predefinied links.
renamed the ch-switzerland flag to just ch to match the rest.
EasyWeatherSetup.php has had its formatting improved some more, and now its CSS is in its own file in the CSS folder. When updating, make sure you download the new easyweathersetup.css file into the css folder.