In development, use with caution.
Adds Nanocharts as a chart option, pulling chart data directly from a MB Nano.
Adds GOES Sat GIFs, currently only avaible in North and South America.
GOES is ready to go, however it's settings code got muddled up in a partial rewrite of the settings page, so would be a PITA to release independant of Nanocharts.
A number of files might show a lot of changes, but that's because I have been working to un-minify a lot of the code.