Added *optional* Google Ananlytics support
Google Analytics script has been added to the main page.
Note that if you leave the Google Analytics ID blank at the bottom of the settings page, the Analytics script will NEVER be present on your webpage.
This allows for site owners to choose to include it, while not changing anything for owners that don't want Google Analytics.